Moved shape factor calculation to utils.
Using a proxy variable `use_sf` instead of `cfg.PARAMS['use_shape_...']`. This variable is set to None, if the acquired key is not found in PARAMS.
